Se añade formulario y controlador de modificación de Servicio

+++ fichajes-web/src/main/java/es/vass/fichajes/utils/ServiceDate.java
@@ -105,7 +105,7 @@ public class ServiceDate {
 			suma += tiempoServicio;
 		}
 		
-		horasTotales = suma/3600000;
+		horasTotales = ((double) suma)/3600000;
 		
 		return horasTotales;
 	}
